requirements.txt
yfinance
pandas
numpy
requests
xlrd
openpyxl
google-generativeai
